Reorganize mocks
Mocks should be easier to find and to modify.
