+++ pkg/slack/filters_test.go
@@ -6,33 +6,6 @@ import (
 	"github.com/ifosch/synthetic/pkg/synthetic"
 )
 
-type MockMessage struct {
-	text    string
-	mention bool
-}
-
-func (m MockMessage) Reply(msg string, inThread bool) {}
-func (m MockMessage) React(reaction string)           {}
-func (m MockMessage) Unreact(reaction string)         {}
-func (m MockMessage) ClearMention() string {
-	return m.Text()
-}
-func (m MockMessage) Thread() bool {
-	return false
-}
-func (m MockMessage) Mention() bool {
-	return m.mention
-}
-func (m MockMessage) Text() string {
-	return m.text
-}
-func (m MockMessage) User() synthetic.User {
-	return nil
-}
-func (m MockMessage) Conversation() synthetic.Conversation {
-	return nil
-}
-
 func TestExactly(t *testing.T) {
 	calls := 0
 	processor := func(synthetic.Message) {
@@ -40,9 +13,9 @@ func TestExactly(t *testing.T) {
 	}
 
 	derivedProcessor := Exactly(processor, "test")
-	derivedProcessor(MockMessage{text: "test"})
-	derivedProcessor(MockMessage{text: "test "})
-	derivedProcessor(MockMessage{text: ""})
+	derivedProcessor(synthetic.NewMockMessage("test", false))
+	derivedProcessor(synthetic.NewMockMessage("test ", false))
+	derivedProcessor(synthetic.NewMockMessage("", false))
 
 	if calls != 1 {
 		t.Logf("Wrong number of executions %v should be 1", calls)
@@ -57,9 +30,9 @@ func TestContains(t *testing.T) {
 	}
 
 	derivedProcessor := Contains(processor, "test")
-	derivedProcessor(MockMessage{text: "test"})
-	derivedProcessor(MockMessage{text: "test "})
-	derivedProcessor(MockMessage{text: ""})
+	derivedProcessor(synthetic.NewMockMessage("test", false))
+	derivedProcessor(synthetic.NewMockMessage("test ", false))
+	derivedProcessor(synthetic.NewMockMessage("", false))
 
 	if calls != 2 {
 		t.Logf("Wrong number of executions %v should be 2", calls)
@@ -74,8 +47,8 @@ func TestMentioned(t *testing.T) {
 	}
 
 	derivedProcessor := Mentioned(processor)
-	derivedProcessor(MockMessage{text: "test", mention: true})
-	derivedProcessor(MockMessage{text: ""})
+	derivedProcessor(synthetic.NewMockMessage("test", true))
+	derivedProcessor(synthetic.NewMockMessage("", false))
 
 	if calls != 1 {
 		t.Logf("Wrong number of executions %v should be 1", calls)
@@ -90,8 +63,8 @@ func TestNotMentioned(t *testing.T) {
 	}
 
 	derivedProcessor := NotMentioned(processor)
-	derivedProcessor(MockMessage{text: "test", mention: true})
-	derivedProcessor(MockMessage{text: ""})
+	derivedProcessor(synthetic.NewMockMessage("test", true))
+	derivedProcessor(synthetic.NewMockMessage("", false))
 
 	if calls != 1 {
 		t.Logf("Wrong number of executions %v should be 1", calls)
