+++ pkg/slack/mocks.go
@@ -9,6 +9,7 @@ type reactionData struct {
 	item     slack.ItemRef
 }
 
+// MockClient is a mocking client for testing.
 type MockClient struct {
 	channels         map[string]*slack.Channel
 	users            map[string]*slack.User
@@ -16,18 +17,22 @@ type MockClient struct {
 	reactionsRemoved []reactionData
 }
 
+// GetConversationInfo returns the channel information for `id`.
 func (c *MockClient) GetConversationInfo(id string, includeLocale bool) (channel *slack.Channel, err error) {
 	return c.channels[id], nil
 }
 
+// GetUserInfo returns the user information for `id`.
 func (c *MockClient) GetUserInfo(id string) (*slack.User, error) {
 	return c.users[id], nil
 }
 
+// NewRTM returns a null Slack RTM.
 func (c *MockClient) NewRTM(options ...slack.RTMOption) *slack.RTM {
 	return nil
 }
 
+// AddReaction registers `reaction` on `item` for validation.
 func (c *MockClient) AddReaction(reaction string, item slack.ItemRef) error {
 	c.reactionsAdded = append(c.reactionsAdded, reactionData{
 		reaction: reaction,
@@ -36,6 +41,7 @@ func (c *MockClient) AddReaction(reaction string, item slack.ItemRef) error {
 	return nil
 }
 
+// RemoveReaction registers `reaction` removal on `item` for validation.
 func (c *MockClient) RemoveReaction(reaction string, item slack.ItemRef) error {
 	c.reactionsRemoved = append(c.reactionsRemoved, reactionData{
 		reaction: reaction,
@@ -89,8 +95,45 @@ func (c *MockClient) reset() {
 	c.reactionsRemoved = []reactionData{}
 }
 
+// NewMockClient creates a new MockClient.
 func NewMockClient() *MockClient {
 	client := &MockClient{}
 	client.reset()
 	return client
+}
+
+// MockRTM is a mocking RTM.
+type MockRTM struct {
+	messagesSent []*slack.OutgoingMessage
+}
+
+// ManageConnection fakes the real Slack RTM connection manager.
+func (rtm *MockRTM) ManageConnection() {}
+
+// NewOutgoingMessage creates a fake message object to send.
+func (rtm *MockRTM) NewOutgoingMessage(text string, channelID string, options ...slack.RTMsgOption) *slack.OutgoingMessage {
+	msg := &slack.OutgoingMessage{
+		Channel: channelID,
+		Text:    text,
+	}
+	for _, option := range options {
+		option(msg)
+	}
+	return msg
+}
+
+// SendMessage fakes sending a message.
+func (rtm *MockRTM) SendMessage(msg *slack.OutgoingMessage) {
+	rtm.messagesSent = append(rtm.messagesSent, msg)
+}
+
+func (rtm *MockRTM) reset() {
+	rtm.messagesSent = []*slack.OutgoingMessage{}
+}
+
+// NewMockRTM creates a new MockRTM.
+func NewMockRTM() *MockRTM {
+	rtm := &MockRTM{}
+	rtm.reset()
+	return rtm
 }
